+++ front/src/utils/MediaQuery.ts
@@ -1,0 +1,3 @@
+export function mediaQueryMaker(minWidth: number, maxWidth: number): string {
+  return `(min-width: ${minWidth}px) and (max-width: ${maxWidth}px)`;
+}
